requirements.txt
africastalking==1.2.4
asgiref==3.5.0
Authlib==1.0.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
contextlib2==21.6.0
coverage==6.3.2
cryptography==37.0.2
Django==3.2.13
django-nose==1.4.7
djangorestframework==3.13.1
idna==3.3
josepy==1.13.0
mozilla-django-oidc==2.0.0
nose==1.3.7
psycopg2==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
schema==0.7.5
sqlparse==0.4.2
urllib3==1.26.9
